03-mcp/hw3/pyproject.toml
[project]
name = "hw3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.14.1",
    "python-dotenv>=1.0.0",
    "requests",
    "minsearch",
]
